perf(scanner): scan monitored domains concurrently and raise the batch
The monitored-domain scanner could not keep its own 14-day cycle. Measured on
production: ~400 scans/day against ~892/day required for the current 12,491
domains, leaving 1,549 domains already overdue. Of the 1,061 domains enrolled
on 2026-07-27, only 182 (17%) had been scanned four days later.

Two causes, from the scan timestamps:
- Each run processed exactly 50 domains (the default getDueDomains limit) at a
  median 6.1s apiece — a 1s politeness sleep plus two HTTP fetches (ads.txt and
  the homepage for language detection), all sequential.
- Runs happened roughly every 2.5-3.5 hours rather than the configured 15
  minutes, consistent with the file's own note that node-cron is unreliable on
  Cloud Run.

This addresses the per-run cost: ads.txt/app-ads.txt fetches now run through a
bounded pool (SCAN_CONCURRENCY, default 8) and the batch is configurable
(SCAN_BATCH_SIZE, default 200). The politeness sleep is dropped because
consecutive items are always different hosts. sellers.json imports stay
sequential, since those files are often hundreds of MB.

Measured against production: 40 domains in 62s (1.55s each) versus 6.1s each
sequentially — a ~4x improvement, which alone lifts throughput to ~1,600/day
even at the current unreliable trigger rate.

Run frequency still needs a Cloud Scheduler trigger against /api/jobs/scan;
that is infrastructure, not code, and is documented separately.

67/67 tests pass.
