+++ backend/src/lib/concurrency.ts
@@ -1,0 +1,18 @@
+/** Run `worker` over `items` with a bounded number of concurrent executions. */
+export async function mapPool<T, R>(
+  items: T[],
+  concurrency: number,
+  worker: (item: T) => Promise<R>,
+): Promise<R[]> {
+  const results: R[] = new Array(items.length);
+  let cursor = 0;
+  async function run(): Promise<void> {
+    while (cursor < items.length) {
+      const idx = cursor++;
+      results[idx] = await worker(items[idx]);
+    }
+  }
+  const runners = Array.from({ length: Math.min(concurrency, items.length) }, run);
+  await Promise.all(runners);
+  return results;
+}
